services/vision-service/requirements.txt
# Vision Service - YOLO + OCR for plate/ID detection
ultralytics>=8.1.0
paddleocr>=2.7.0
paddlepaddle>=2.6.0
opencv-python-headless>=4.9.0
numpy>=1.26.0
httpx>=0.26.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pillow>=10.2.0
